More clearly dodge responsibility
The whole mess that is key configuration discovery needs real work.
There are cases where trust suffices, but we can't rely on that here.
What this does is tweak the language to more clearly note the lines of
responsibility for this draft.

I also noted that we didn't talk about trial decryption and tracking, so
I added that.  It seemed like we didn't need to mention multiple keys
with the same identifier twice after that.