Proxies don't need TLS to the server
But they should.  Consider the case where clients have equal-sized
messages, but are making requests using different key identifiers or
different request targets.  If that information could be linked to
certain clients, encryption would increase the size of the anonymity set
from the perspective of a passive observer for little extra cost.

Closes #56.