A start on operation and deployment stuff
I wanted to add more then just interception text here.  So I added a
little about performance and the ODoH study made some of those costs
clear enough to cite.

I anticipate this section growing over time.  Some of the material we
might add is already in other sections of the doc (DoS management at the
proxy, for example), but we might move that around some.  I also expect
the working group to focus rather narrowly on this topic.