Point in the direction of arbitrary forwarding
Truly, the most interesting forwarding is what occurs at the Oblivious
Request Resource, but this is a point of inflexibility that some people
struggle with.  No point in hiding how we might intend for a generic
proxy to operate.  The alternatives people invent are likely to be
worse if we don't at least point in the rough direction of a good
solution.

Closes #105.