Bind HPKE info to the configuration
We previously put the details from the key configuration (key id, KEM,
KDF, AEAD) in the AAD for the request.  But @davidben noted that this
meant that this wasn't bound to the response.

Analysis shows that this isn't a security problem, but while that
depends on some very reasonable assumptions about how keys are used,
there is no reason not to do better.

Performance-wise, this change is a wash.  The AAD input is now empty,
which should make this a tiny bit faster to execute.  The downside is
that the info is no longer static, which means that it is a tiny bit
slower, especially on the server side where it is harder to just pass a
constant value into SetupBaseR and the bits from the header into the
AAD.  But those are extraordinarily trivial differences.

The spec doesn't really change much either, though I've taken this
chance to be a bit more explicit about what functions in HPKE are run
and what their inputs are.

I will do examples in a follow-up so that you can see the effect on
those (the numbers change, that's all).

Code changes are fairly simple also, see
https://github.com/martinthomson/ohttp/pull/21 for an example.  Note
that the biggest part of this is down to a structural choice, the actual
changes are minor.  Response handling doesn't change at all.

Closes #131.