Rewrite privacy text
I was just going to add something about correlating requests based on
their content, but then it seemed like the text was duplicated some.

Here is my attempt at making this better.  I also expanded the notion of
"configuration" to include the identity of the relay.  That made it a
little tricky - we really do want diversity of relays - but it makes
more sense to talk about minimum anonymity set size rather than to imply
that there can be a single configuration.

Closes #149.