Don't prohibit cookies
Not even once, down from two.

Using cookies is dumb, but we don't need to ban them so directly.  There
are tons of ways in which information from a previous request might leak
across into the current one.  The point is to highlight the risk of
using them.

A direct prohibition doesn't make sense if the use of the cookie is
well understood (not that attaining such an understanding is likely to
be easy).

Closes #160.