Use the plural
We needed two media types because the payload doesn't distinguish
between them, but that isn't clear.

It is possible to define a signal to distinguish request from response
(like message/bhttp does), so I don't want to definitively say "two"
media types.  Of course, if you are reusing this specification, you
might not get that.