More on replacing the message types for OHTTP
I think that this is more responsive to Tommy's request for changes.

This adds more text on what might be done to continue to use OHTTP but
with a different underlying message representation (other than
message/bhttp) with the same encapsulation method; OR a completely
different encapsulation method, at which point the message
representation matters less.

I've moved the repurposing section up adjacent to this text as a result.
There are a couple of tiny changes in that text so that it doesn't
self-reference and so that it fits better into context.

cc @tfpauly so he sees this note.