Better describe the content of various things
We previously only had figures, which are sort of illustrative only,
they don't really work as normative specification.
The content of these really needs to have words that describe what was
intended.

The definition for the request was already sort-of there, so what I've
added is a tiny bit repetitive, but I think that it's OK in this case to
have the descriptions at different levels of specificity.

The big error here was that the length field in the key configuration
was impossible to decipher.  We never said whether it was length in
bytes (what has been implemented) or length in number of items.  The
bounds on the field length were not helpful here as they were not
correct in any way.  I've corrected this value, but would appreciate a
sanity check.

No one identified this as a problem in the reviews.  I guess that means
that the QUIC format is clear and understandable.  Maybe I should have
made it less so.  The TLS format is consistently misunderstood, which I
used to think was a disadvantage...