A relay can pad too
This came out of #257, where we realized that the relay padding is
useful, but needs to employ different methods than client or gateway.
The relay can't use binary HTTP padding; it has to insert stuff at the
HTTP layer.  Include citations for padding in those documents.

Note that these citations are slightly misleading.  Most relays won't
have access at the level necessary to add that padding.  More likely,
the relay will need to do as @chris-wood originally suggested, namely
adding fields of an appropriate size.  But we don't have good citations
for that, so I'm cheaping out.  (Note that RFC 9112, Section 7.1.1 talks
about using chunk extensions for padding, but I'm going to pretend that
doesn't exist; it's even less usable than the HTTP/2 and HTTP/3
options.)