Include best advice on setting anti-replay windows
This really is pretty hard.  I spent a lot of time on this when
determining anti-replay in our TLS implementation.  Based on numbers
informally provided by Meta, I concluded that 10s is probably OK in that
context.  The constraints are different here though.  Rejection of
requests here is likely a lot more disruptive, so a larger window might
be needed.  On the other hand, an insistence on having better clocks is
not entirely unreasonable.

Closes #261.