Separate target and gateway
This talks about the odd case where target and gateway are separate.
The primary insight here is that the target is just a regular HTTP
resource and the requests from the gateway aren't special.  Problems
only occur if the gateway is privileged in some way.

If abuse comes via the gateway, then the target would ordinary block the
client, which tends to affect a bunch of bystanders in this model.  So
there is pressure to do something special for the gateway.  In that
case, the target is trusting the gateway in some way.  Explain that.

The heading here is the only case where we use the short names of the
entities, but I can't stand short headings.